feat: get blog working
