+++ src/components/content/blog/BlogCard.tsx
@@ -1,11 +1,13 @@
 import clsx from 'clsx';
 import { format } from 'date-fns';
 import * as React from 'react';
+// import { useEffect, useState } from 'react';
 import { HiOutlineClock, HiOutlineEye } from 'react-icons/hi';
 
 import Accent from '@/components/Accent';
 import Tag from '@/components/content/Tag';
 import CloudinaryImg from '@/components/images/CloudinaryImg';
+import Img from '@/components/images/Img';
 import UnstyledLink from '@/components/links/UnstyledLink';
 
 import { BlogFrontmatter, InjectedMeta } from '@/types/frontmatters';
@@ -21,6 +23,36 @@ export default function BlogCard({
   checkTagged,
   onClick,
 }: BlogCardProps) {
+  const non_cloudinary = post.banner.startsWith('/folio-v2/');
+  let img;
+  if (non_cloudinary) {
+    img = (
+      <Img
+        noStyle
+        className='pointer-events-none overflow-hidden rounded-t-md'
+        publicId={post.banner}
+        alt='Photo taken by me'
+        width={1200}
+        height={(1200 * 2) / 5}
+        aspect={{ height: 2, width: 5 }}
+        preview={false}
+      />
+    );
+  } else {
+    img = (
+      <CloudinaryImg
+        noStyle
+        className='pointer-events-none overflow-hidden rounded-t-md'
+        publicId={`theodorusclarence/banner/${post.banner}`}
+        alt='Photo taken from unsplash'
+        width={1200}
+        height={(1200 * 2) / 5}
+        aspect={{ height: 2, width: 5 }}
+        preview={false}
+      />
+    );
+  }
+
   return (
     <li
       className={clsx(
@@ -38,16 +70,7 @@ export default function BlogCard({
         href={`/blog/${post.slug}`}
       >
         <div className='relative'>
-          <CloudinaryImg
-            noStyle
-            className='pointer-events-none overflow-hidden rounded-t-md'
-            publicId={`theodorusclarence/banner/${post.banner}`}
-            alt='Photo taken from unsplash'
-            width={1200}
-            height={(1200 * 2) / 5}
-            aspect={{ height: 2, width: 5 }}
-            preview={false}
-          />
+          {img}
           <div
             className={clsx(
               'absolute bottom-0 w-full px-4 py-2',
