+++ src/pages/projects/[...slug].tsx
@@ -17,6 +17,7 @@ import TableOfContents, {
   HeadingScrollSpy,
 } from '@/components/content/TableOfContents';
 import CloudinaryImg from '@/components/images/CloudinaryImg';
+// import Img from '@/components/images/Img';
 import Layout from '@/components/layout/Layout';
 import CustomLink from '@/components/links/CustomLink';
 import Seo from '@/components/Seo';
